#81 Cleanup code
